Remove abstract from setUpListenerOnWebViewChanges
