+++ diagram-viewer/src/main/java/com/powsybl/diagram/viewer/common/AbstractDiagramController.java
@@ -10,10 +10,12 @@ package com.powsybl.diagram.viewer.common;
 import com.google.common.io.ByteStreams;
 import com.powsybl.iidm.network.*;
 import javafx.beans.property.StringProperty;
+import javafx.concurrent.Worker;
 import javafx.fxml.FXML;
 import javafx.scene.control.TextArea;
 import javafx.scene.input.ScrollEvent;
 import javafx.scene.web.WebView;
+import netscape.javascript.JSObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -40,7 +42,19 @@ public abstract class AbstractDiagramController {
     private String html;
     private String js;
 
-    protected abstract void setUpListenerOnWebViewChanges(java.lang.Object jsHandler);
+    protected void setUpListenerOnWebViewChanges(java.lang.Object jsHandler) {
+        // Set up the listener on WebView changes
+        // A listener has to be added as loading takes time - execute once the content is successfully loaded
+        diagramWebView.getEngine().getLoadWorker().stateProperty().addListener((observable, oldValue, newValue) -> {
+            if (Worker.State.SUCCEEDED == newValue) {
+                JSObject window = (JSObject) diagramWebView.getEngine().executeScript("window");
+                window.setMember("jsHandler", jsHandler);
+                // For easier debugging, redirect the console.log to the jsHandler
+                diagramWebView.getEngine().executeScript("console.log = function(message) " +
+                        "{ jsHandler.log(message); };");
+            }
+        });
+    }
 
     protected void init() throws IOException {
         // Add Zoom management
